Fix #50 & add #60 (disabled options)
I updated the template to switch back to default Bootstrap dropdown
markup. Because of this, the scrollbar covers up the rounded corners on
Windows, which doesn't look quite as good but is worth it to be
consistent. I removed the extra CSS that is no longer needed. Also added
support for disabled options (``<option disabled>Mustard</option>``).