Fix #65. Update dividers
I updated the dividers used for the optgroups to now use the default
Bootstrap CSS class. I've also included the ``data-divider`` option. Set
``data-divider="true"`` to turn an ``option`` into a divider.

I simplified ``data-size="auto"`` - it now opens to be as tall as
possible, but with a 14px padding between the edge of the menu and the
edge of the window. Setting a number for ``data-size`` still results in
similar functionality.

I also set a min-height on the options to take care of #65 as it was
much cleaner than a javascript solution.

I also updated the reloadLi function so if an option set to selected is
appended to the select, the selectpicker text is updated to reflect the
new selected option.