Fix #155 (properly)
The new changes fixed #165, but broke #155 again. The issue was occurring in Firefox when a select already had an option selected by default, selected="selected". This was causing issues with `prevValue` and thus a change event wasn't occurring when it should have been. This now only performs `removeAttr()` on items items that are already set to `selected="selected` which for some reason works.