Fix #194 & #130 (problems with size)
setSize() is now called when the select is clicked, instead of on
init(). This fixes issues with hidden selects/data-size in general. I've
also added a container to the menu, which doesn't really do anything
now, but it lays the framework for when we add a "select all/deselect
all" feature, as well as a filter option (coming soon).