Add "fit" width option. Other width changes
"fit" is added as an option for width to dynamically change the width of
the select to whatever is currently selected (adds fit-width class).
It's now possible to change the width programmatically as well, by
removing inline CSS that's added by other options. Change default width
value from null to false, to be more in line with other default option
values.