Update bootstrap-select.min.js
Change in displaying and searching grouped elements. Now when some element belongs to some group the name of that group will be displayed as well as that element. This is especially useful when there are few elements with the same name in many categories - You can always see the category (group) name.
The same as in bootstrap-select.js