Fix active toggle for <li> tags
The `$items` refers to a list of elements with the following hierarchy:

    <li class="active">
      <a class="active">option 1</a>
    </li>
    <li class="">
      <a class="">option 2</a>
    </li>
    <li class="">
      <a class="">option 3</a>
    </li>

When the up and down arrows are pressed the code attempted to toggle the
active class. Bootstrap in turn highlighted the item. However, Bootstrap
only looks at the <li> tag with the active class not the <a> (which is
ignored by the CSS).

This patch instead loops over the list of <a> tags like it did before
but also toggles the active class on each parent. This is essentially
the same logic as before but with the added bonus of each item's parent
getting the active or not active state.

The result is that pressing the up and down arrows now work in
Bootstrap 3 as expected.