improve performance when opening select
The new method allows the selects to load faster on init, but increased
rendered time when first opening the select (liHeight was moved from
init to setSize). This slowness is really only noticeable on a select
with a lot of options (#791), but this new method drastically decreases
the time taken to open the menu (from ~100ms to ~25ms for a select with
1000 options). Opening the menu after that reduces to around ~5ms as the
size calculations are cached.