fix single select title functionality. improve speed
Don't select the blank title option of the select if there is already an
option that has the "selected" attribute. Previously, if a single select
had a title, it would override the option with the "selected" attribute
and select the blank title option.

Also, slight improvement in rendering speed for selects with a lot of
options by using native JS.