trigger setSize on refresh event
setSize needs to be executed in order to reflect new size when the options have been populated programmatically (eg. via ajax). This issue happens when listening for **show.bs.select** and populate the options within that method.

Test case:
```javasript
$('.selectpicker').on('show.bs.select', function (e) {
        // do ajax request and when done,
        $('.selectpicker').selectpicker('refresh'); // needs to trigger setSize
});
```

Currently when refresh event is triggered it opens the dropdown without size, it shows all items, and once closed and opened again it shows the correct size. This is because setSize is set only when 
```javascript
this.$button.on('click', function () {
        that.setSize();
});
```