Fix <li> tags inside <option> being treated as options
The findLis() function used a too broad selector which included <li>
tags inside <option>. For example for this <select>:

    <select>
      <option><ul><li>demo</li></option>
    </select>

findLis() would find two <li>s in generated code: one for <option> and one
inside it.

One side effect of this bug is that check mark is shown properly only for
the first option. For following options the `selected` class is assigned
to incorrect <li>s.