only prepend titleOption when necessary. Don't set innerHTML for title option. Fix #26
This fixes a validation issue in Firefox. Appending/prepending options to the select cause the :valid pseudo class to be applied incorrectly. This is still occurring when prepending an option in Firefox, but as this appears to be a bug with Firefox, I can't see a way to workaround it.