Fixes two `TypeError` exceptions when up/down arrow pressed (#2585)
* fixes `Cannot read property 'position' of undefined` error.

when using live search, if a user enters a value that does not return
any results and then presses the up/down arrows, this error occurs. this
commit fixes that error.

* fixes `Cannot read property 'index' of undefined` error.

when using live search, if a user enters a value that does not return
any results and then presses the up/down arrows, this error occurs. this
commit fixes that error.

Co-authored-by: RJ Regenold <rj@providerscience.com>