Fixes: Opening dropdown does not scroll to selected option first time, but does second time
Fix for issue:

The dropdown renders with an option selected. But when I click to open the dropdown, it does not show the selected option, but instead the list is at the top. However, close the dropdown and click to open it again and this time it will be scrolled to the selected option. #2838